test delegated deposit addresses
