+++ GasProfile.json
@@ -17,10 +17,10 @@
   "burnMicroDollar": 42704,
   "reduceToBurnWithChange": 42768,
   "emptyToBurnWithChange": 42768,
-  "reduceApprovalReducingToBurn": 49514,
-  "emptyApprovalReducingToBurn": 34589,
-  "reduceApprovalEmptyingToBurn": 34525,
-  "emptyApprovalEmptyingToBurn": 24673,
+  "reduceApprovalReducingToBurn": 49521,
+  "emptyApprovalReducingToBurn": 34596,
+  "reduceApprovalEmptyingToBurn": 34532,
+  "emptyApprovalEmptyingToBurn": 24676,
   "reduceToNewWithRefund": 35415,
   "reduceToExistingWithRefund": 27551,
   "emptyToExistingWithRefund": 24661,
@@ -39,10 +39,10 @@
   "burnMicroDollarWithRefund": 28982,
   "reduceToBurnWithChangeWithRefund": 29014,
   "emptyToBurnWithChangeWithRefund": 29014,
-  "reduceApprovalReducingToBurnWithRefund": 32746,
-  "emptyApprovalReducingToBurnWithRefund": 29917,
-  "reduceApprovalEmptyingToBurnWithRefund": 29885,
-  "emptyApprovalEmptyingToBurnWithRefund": 24673,
+  "reduceApprovalReducingToBurnWithRefund": 32749,
+  "emptyApprovalReducingToBurnWithRefund": 29921,
+  "reduceApprovalEmptyingToBurnWithRefund": 29889,
+  "emptyApprovalEmptyingToBurnWithRefund": 24676,
   "sponsorGas": 208304,
   "sponsorGas2": 203365
 }
